Moved toolbar to fragment to allow hide-on-scroll
