+++ app/src/main/res/layout/activity_list_view.xml
@@ -15,97 +15,12 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program.  If not, see <http://www.gnu.org/licenses/>.
   -->
-
-<androidx.coordinatorlayout.widget.CoordinatorLayout
+<FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_list_view"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:id="@+id/appBar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        app:elevation="0dp">
-
-        <com.google.android.material.appbar.CollapsingToolbarLayout
-            android:id="@+id/toolbar_layout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:fitsSystemWindows="true"
-            app:contentScrim="?attr/colorPrimary"
-            app:layout_scrollFlags="scroll|enterAlways"
-            app:toolbarId="@+id/toolbar" >
-
-            <androidx.appcompat.widget.Toolbar
-                android:id="@+id/toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:visibility="visible"
-                app:contentInsetStartWithNavigation="0dp"
-                app:titleMarginStart="0dp"
-                tools:title="@string/app_name">
-
-                <androidx.appcompat.widget.SearchView
-                    android:id="@+id/search_view"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content">
-                </androidx.appcompat.widget.SearchView>
-
-            </androidx.appcompat.widget.Toolbar>
-
-            <com.google.android.material.card.MaterialCardView
-                android:id="@+id/home_toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="@dimen/spacer_2x"
-                android:layout_marginTop="@dimen/spacer_1hx"
-                android:layout_marginEnd="@dimen/spacer_2x"
-                android:layout_marginBottom="@dimen/spacer_1hx"
-                app:cardCornerRadius="30dp"
-                app:cardElevation="2dp"
-                app:strokeWidth="0dp">
-
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:orientation="horizontal">
-
-                    <androidx.appcompat.widget.AppCompatImageButton
-                        android:id="@+id/menu_button"
-                        android:layout_width="?android:attr/actionBarSize"
-                        android:layout_height="?android:attr/actionBarSize"
-                        android:background="?android:attr/selectableItemBackgroundBorderless"
-                        android:layout_gravity="center_vertical|start"
-                        android:paddingStart="@dimen/spacer_1x"
-                        android:paddingTop="@dimen/spacer_2x"
-                        android:paddingEnd="@dimen/spacer_1x"
-                        android:paddingBottom="@dimen/spacer_2x"
-                        android:tint="@color/text_color"
-                        android:src="@drawable/ic_menu_grey"
-                        android:contentDescription="@string/menu"/>
-                    <androidx.appcompat.widget.AppCompatTextView
-                        android:id="@+id/search_text"
-                        android:layout_width="0dp"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center_vertical"
-                        android:layout_marginStart="@dimen/spacer_1x"
-                        android:layout_marginEnd="@dimen/spacer_1x"
-                        android:layout_weight="1"
-                        android:ellipsize="end"
-                        android:gravity="start"
-                        android:lines="1"
-                        android:textSize="16sp"
-                        android:text="@string/search_in_all"/>
-                </LinearLayout>
-
-            </com.google.android.material.card.MaterialCardView>
-
-        </com.google.android.material.appbar.CollapsingToolbarLayout>
-
-    </com.google.android.material.appbar.AppBarLayout>
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container"
@@ -154,4 +69,4 @@
         app:tint="@color/white"
         android:contentDescription="@string/open_fab"/>
 
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
+</FrameLayout>
